+++ src/features/expense/model/useCategoryStore.ts
@@ -1,0 +1,10 @@
+import { create } from 'zustand';
+import { Category, CategoryState } from './types';
+
+const initialCategoryState: CategoryState = {
+  categories: [],
+};
+
+const useCategoryStore = create<CategoryState>((set) => ({
+  ...initialCategoryState,
+}));
