+++ src/features/expense/model/useExpenseFormStore.ts
@@ -1,0 +1,34 @@
+import { create } from 'zustand';
+import { Expense, ExpenseFormActions } from './types/Expense';
+
+const initialExpenseFormState: Omit<Expense, 'id'> = {
+  date: new Date(),
+  memo: '',
+  amount: 0,
+};
+
+const useExpenseFormStore = create<Omit<Expense, 'id'> & ExpenseFormActions>(
+  (set) => ({
+    ...initialExpenseFormState,
+    setDate: (value) => {
+      set({ date: value });
+    },
+    setProvidedCategory: (value) => {
+      set({ providedCategory: value });
+    },
+    setCategory: (value) => {
+      set({ category: value });
+    },
+    setMemo: (value) => {
+      set({ memo: value });
+    },
+    setAmount: (value) => {
+      set({ amount: value });
+    },
+    resetForm: () => {
+      set({ ...initialExpenseFormState });
+    },
+  })
+);
+
+export default useExpenseFormStore;
